+++ bin/gitm-undo.ts
@@ -9,9 +9,14 @@ const { queue } = require('./core/queue')
 const {
     getIsGitProject,
     getGitLogs,
-    getGitLogsByCommitIDs
+    getGitLogsByCommitIDs,
+    getCurrentBranch
 } = require('./core/git/index')
-const { getRevertCache, addRevertCache } = require('./core/cache/index')
+const {
+    getRevertCache,
+    addRevertCache,
+    setRevertCache
+} = require('./core/cache/index')
 const { createArgs, echo } = require('./core/utils/index')
 if (!getIsGitProject()) {
     sh.echo(red('当前目录不是git项目目录'))
@@ -31,6 +36,8 @@ interface GitmBuildOption {
     merges: boolean
     lastet: string
     limit: number
+    calc: boolean
+    calcAll: boolean
 }
 
 /**
@@ -42,13 +49,16 @@ interface GitmBuildOption {
  */
 function getRevertCommitIDs(commitIDs: string[]): string[] {
     const revertCache = getRevertCache()
+    const current = getCurrentBranch()
     let len = commitIDs.length
     while (len--) {
         const _before = revertCache.findIndex(
-            (item: RevertCacheType) => item.before['%H'] === commitIDs[len]
+            (item: RevertCacheType) =>
+                item.branch === current && item.before['%H'] === commitIDs[len]
         )
         const _after = revertCache.findIndex(
-            (item: RevertCacheType) => item.after['%H'] === commitIDs[len]
+            (item: RevertCacheType) =>
+                item.branch === current && item.after['%H'] === commitIDs[len]
         )
         if (_before > -1 || _after > -1) {
             echo(
@@ -65,6 +75,71 @@ function getRevertCommitIDs(commitIDs: string[]): string[] {
 }
 
 /**
+ * 清理撤销失败的记录
+ * 1. 执行冲突导致失败，after = null
+ * 2. 没有找到撤销记录
+ * 3. 有撤销记录，但已经恢复
+ *
+ * @param all - 是否清理全部，默认只清理当前分支
+ */
+function calculate(all = false) {
+    const keys = ['%H', '%T', '%P', '%aI', '%an', '%s', '%b']
+    const revertCache = getRevertCache()
+    const current = getCurrentBranch()
+    let len = revertCache.length
+    while (len--) {
+        let { before, after, branch } = revertCache[len],
+            _undoLogs = [],
+            _redoLogs = []
+        if (!after) {
+            // 没有after，尝试恢复
+            const _logs = getGitLogs({
+                lastet: '1d',
+                limit: 40,
+                noMerges: true,
+                keys,
+                branch,
+                grep: before['%H']
+            })
+            if (_logs && _logs.length > 0) revertCache[len].after = _logs[0]
+        }
+        _undoLogs = getGitLogs({
+            lastet: '60d',
+            limit: 500,
+            noMerges: true,
+            keys,
+            branch,
+            grep: before['%H']
+        })
+        if (after) {
+            _redoLogs = getGitLogs({
+                lastet: '60d',
+                limit: 500,
+                noMerges: true,
+                keys,
+                branch,
+                grep: after['%H']
+            })
+        }
+		console.log(1000, _undoLogs, _redoLogs)
+        // 不检测全部分支时，跳过非当前分支
+        if (!all && current !== branch) continue
+        // 没有找到该记录的撤销记录 或 撤销的记录被恢复过
+        if (_undoLogs.length === 0 || _redoLogs.length > 0) {
+            echo(
+                yellow(
+                    _undoLogs.length === 0
+                        ? `检测到 ${revertCache[len].before['%H']} 这条记录撤销失败，已删除相关日志`
+                        : `检测到 ${revertCache[len].before['%H']} 这条记录已经被恢复了，已删除相关日志`
+                )
+            )
+            revertCache.splice(len, 1)
+        }
+    }
+    setRevertCache(revertCache)
+}
+
+/**
  * gitm undo
  */
 program
@@ -78,12 +153,15 @@ options.forEach((o: GitmarsOptionOptionsType) => {
     program.option(o.flags, o.description, o.defaultValue)
 })
 // .arguments('[commitid...]')
+// .option('--calc', '清理当前分支撤销失败的记录')
+// .option('--calcAll', '清理所有撤销失败的记录')
 // .option('--no-merges', '是否排除merge的日志')
 // .option('-m, --mode [mode]', '针对撤销一次merge记录，需要传入类型：1 = 保留当前分支代码，2 = 保留传入代码', 1)
 // .option('--lastet [lastet]', '查询在某个时间之后的日志，填写格式：10s/2m/2h/3d/4M/5y', '7d')
 // .option('--limit [limit]', '最多查询的日志条数', 20)
 program.action(async (commitid: string[], opt: GitmBuildOption) => {
     const keys = ['%H', '%T', '%P', '%aI', '%an', '%s', '%b'] as const
+    const current = getCurrentBranch()
     let logList: Array<
             {
                 [key in typeof keys[number]]: string
@@ -92,6 +170,14 @@ program.action(async (commitid: string[], opt: GitmBuildOption) => {
         cmd: Array<CommandType | string> = [],
         commitIDs: string[] = [], // 需要执行的commitID
         mode = ''
+    if (opt.calc) {
+        calculate()
+        return
+    }
+    if (opt.calcAll) {
+        calculate(true)
+        return
+    }
     if (!opt.limit) opt.limit = 20
     if (opt.mode) mode = ' -m ' + Math.abs(Number(opt.mode))
     if (commitid.length > 0) {
@@ -158,22 +244,34 @@ program.action(async (commitid: string[], opt: GitmBuildOption) => {
             fail: '出错了，请根据提示处理'
         }
     }))
-    // 执行
-    queue(cmd).then(() => {
-        const revertCacheList = logList.map(log => {
-            const cache = { before: log, after: null }
-            const _logs = getGitLogs({
-                lastet: '1d',
-                limit: opt.limit * 2,
-                noMerges: true,
-                keys,
-                grep: log['%H']
-            })
-            if (_logs.length > 0) cache.after = _logs[0]
-            return cache
+    // 先保存缓存
+    const revertCacheList = logList.map(log => {
+        const cache = { before: log, after: null, branch: current }
+        const _logs = getGitLogs({
+            lastet: '1d',
+            limit: opt.limit * 2,
+            noMerges: true,
+            keys,
+            branch: current,
+            grep: log['%H']
         })
-        addRevertCache(revertCacheList)
+        console.log(_logs, logList, log)
+        if (_logs.length > 0) cache.after = _logs[0]
+        return cache
     })
+    addRevertCache(revertCacheList)
+    // 执行
+    queue(cmd)
+        .then(() => {
+            calculate()
+        })
+        .catch(() => {
+            echo(
+                yellow(
+                    '处理冲突之后，执行：gitm undo --calc'
+                )
+            )
+        })
 })
 
 program.parse(process.argv)