+++ docs/changelog.md
@@ -8,7 +8,7 @@
 2. 新增`undo`指令，用于撤回代码
 3. 新增`redo`指令，用于恢复撤销的代码
 4. `cleanbranch`指令流程优化，预览之后可选择分支进行清理
-5. `undo`指令新增`--no-merges` `--limit` `--lastet`传参
+5. `undo`指令新增`--no-merges` `--limit` `--lastet` `--calc` `--calcAll`传参
 6. `log`指令增加`--no-merges`传参，用于排除merge记录
 7. `process.exit`替换`sh.exit`